feat: synchronizing the remote Run folder in remote system

+++ front_end/openVRE/public/phplib/classes/RemoteSSH.php
@@ -249,7 +249,7 @@ class RemoteSSH {
         }
     }
 
-    public function executeRsyncCommandForWorkingDir($sshCredentials, $localDir, $remoteDir, $server)
+    public function executeRsyncCommandForWorkingDir($sshCredentials, $localDir, $remoteDir, $server, $singularityImage = null)
     {
         $sshPrivateKey = trim($sshCredentials['private_key']);
         $username = $sshCredentials['username'];
@@ -280,6 +280,16 @@ class RemoteSSH {
                 }
                 $_SESSION['errorData']['Info'][] = "Created remote dir: $remoteDir";
             }
+            // Check for Singularity Image
+            error_log("Looking for Singularity image in path: $singularityImage");
+            $checkSifCommand = "[ -f \"$singularityImage\" ] && echo 'SIFExists' || echo 'SIFMissing'";
+            $sifStatus = trim($ssh->exec($checkSifCommand));
+            $sifStatus = preg_match('/(SIFExists|SIFMissing)/', $sifStatus, $matches) ? $matches[1] : "Unknown";
+            if ($sifStatus !== "SIFExists") {
+                $_SESSION['errorData']['Error'][] = "Required Singularity image is missing: $singularityImage";
+                return false;
+            }
+            $_SESSION['errorData']['Info'][] = "Singularity image found: $singularityImage";
             // Perform rsync
             $tempKeyFile = tempnam(sys_get_temp_dir(), 'ssh_key_');
             file_put_contents($tempKeyFile, $formattedKey);
